+++ core/blocks.py
@@ -1,5 +1,7 @@
+from django.utils.html import format_html
 from wagtail.core import blocks
 from wagtail.images.blocks import ImageChooserBlock
+from wagtailmedia.blocks import AbstractMediaChooserBlock
 
 
 class LinkBlock(blocks.StructBlock):
@@ -19,3 +21,30 @@ class LinkWithSourceBlock(LinkBlock):
 class LinkWithImageAndContentBlock(LinkWithSourceBlock):
     content = blocks.RichTextBlock()
     image = ImageChooserBlock()
+
+
+class MediaChooserBlock(AbstractMediaChooserBlock):
+    def render_basic(self, value, context=None):
+        if not value:
+            return ''
+        return value.file.url
+
+
+class VideoBlock(blocks.StructBlock):
+    width = blocks.IntegerBlock()
+    height = blocks.IntegerBlock()
+    video = MediaChooserBlock()
+
+    def render(self, value, context=None):
+        if not value:
+            return ''
+        return format_html(
+            f"""
+                    <div>
+                        <video width="{self.width}" height="{self.height}" controls>
+                            {source}
+                            Your browser does not support the video tag.
+                        </video>
+                    </div>
+                    """
+        )